centos & freebsd compatibility

+++ src/main.c
@@ -734,7 +734,7 @@ int main(int argc, char **argv) {
   // do ajax script
   if (globaldata.gd_semaphore && globaldata.gd_semaphore->sem_keep)
     printf("<script type=\"text/javascript\">do_ajax('client',%d,'%s',new Date().getTime())</script>\n",
-          globaldata.gd_semaphore->sem_key,
+          (int)globaldata.gd_semaphore->sem_key,
           getenv("SCRIPT_NAME"));
 
   printf(HTM_FOOTER);
